Make composer-shim POSIX-compatible
This shim file has `#!/bin/sh` declared as its shebang. However, it used a
double bracket operator internally, which is a bashism and does not work in
all shells that `/bin/sh` may represent.

AFAIU, the new version of the conditional should be equivalent but POSIX-
compatible, and therefore work in all shells.