Fix in-place extension enabling on PHP7.2 upwards
PHP7.2 allows the extension name <ext> to be used rather than the dll
name `php_<ext>.dll`, as used in the extensions directory. Previously
we used the dll name, and because we couldn't find a match to the new
format the enabled extension was added at the end of the php.ini file.
It is now added in-place.